Updated drag and drop and resize to grab the data neccessary to update state. Will update store/reducer next
