+++ Frontend/src/Actions/dragAndDrop.js
@@ -8,28 +8,6 @@ const draggableOptions = {
     },
     // enable autoScroll
     autoScroll: true,
-
-  onmove: event => {
-    const target = event.target
-    target.style.zIndex = '1';
-  // keep the dragged position in the data-x/data-y attributes
-  const x = (parseFloat(target.getAttribute('data-x')) || 0) + event.dx
-  const y = (parseFloat(target.getAttribute('data-y')) || 0) + event.dy
-
-  // translate the element
-  target.style.webkitTransform =
-  target.style.transform =
-    'translate(' + x + 'px, ' + y + 'px)'
-
-  // update the posiion attributes
-  target.setAttribute('data-x', x);
-  target.setAttribute('data-y', y);
-  },
-  onend: event => {
-    const target = event.target
-    target.style.zIndex = '0';
-
-  }
 }
 
 export default draggableOptions
